Made suggested changes.
As recommended by the scan changed the main div to <main> and the div.attribution into footer thus gaining the contentinfo landmark.

It also recommended adding an h1 but that is to large.

My solution is slightly larger when site did comparison but my width is fitting the style measurements provided so leaving it as is.